requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
configparser==3.5.0
dnspython==1.15.0
geoip2==2.5.0
IPy==0.83
maxminddb==1.3.0
rblwatch==0.3.0
requests==2.14.2
urllib3==1.20
